+++ Editor/Drawers/SaintsDictionary/DictionaryWrap/SaintsDictionaryWrapDrawerIMGUI.cs
@@ -1,0 +1,7 @@
+namespace SaintsField.Editor.Drawers.SaintsDictionary.DictionaryWrap
+{
+    public partial class SaintsDictionaryWrapDrawer
+    {
+        protected override bool UseCreateFieldIMGUI => true;
+    }
+}