+++ Editor/Drawers/SaintsDictionary/DictionaryWrap/SaintsDictionaryWrapDrawerUIToolkit.cs
@@ -1,0 +1,43 @@
+using System.Collections.Generic;
+using System.Reflection;
+using SaintsField.Editor.Drawers.SaintsRowDrawer;
+using SaintsField.Editor.Utils;
+using SaintsField.Interfaces;
+using UnityEditor;
+using UnityEngine;
+using UnityEngine.UIElements;
+
+namespace SaintsField.Editor.Drawers.SaintsDictionary.DictionaryWrap
+{
+    public partial class SaintsDictionaryWrapDrawer
+    {
+        protected override bool UseCreateFieldUIToolKit => true;
+
+        protected override VisualElement CreateFieldUIToolKit(SerializedProperty property, ISaintsAttribute saintsAttribute,
+            IReadOnlyList<PropertyAttribute> allAttributes, VisualElement container, FieldInfo info, object parent)
+        {
+            // string label = GetPreferredLabel(property);
+            //
+            SerializedProperty realProp = property.FindPropertyRelative("value") ?? SerializedUtils.FindPropertyByAutoPropertyName(property, "value");
+            Debug.Assert(realProp != null, property.propertyPath);
+
+            FieldInfo realInfo = ReflectUtils.GetElementType(info.FieldType).GetField("value", BindingFlags.Public | BindingFlags.Instance | BindingFlags.FlattenHierarchy);
+
+            InHorizontalLayout = true;
+
+            // Debug.Log($"{property.propertyPath}: {string.Join(",", allAttributes)}");
+
+            (PropertyAttribute[] _, object realParent) = SerializedUtils.GetAttributesAndDirectParent<PropertyAttribute>(realProp);
+
+            // UnityFallbackUIToolkit(FieldInfo info, SerializedProperty property, IReadOnlyList<PropertyAttribute> allAttributes, VisualElement containerElement, string passedPreferredLabel, IReadOnlyList<SaintsPropertyInfo> saintsPropertyDrawers, object parent)
+
+            using(new SaintsRowAttributeDrawer.ForceInlineScoop(true))
+            {
+                return UnityFallbackUIToolkit(
+                    realInfo, realProp, allAttributes, container, GetPreferredLabel(property), SaintsPropertyDrawers,
+                    realParent);
+            }
+            // return PropertyFieldFallbackUIToolkit(realProp);
+        }
+    }
+}